+++ R/visualization.R
@@ -22,9 +22,14 @@ NULL
 #' @param ncol Number of columns to plot
 #' @param fast If true, use \code{image} to generate plots; faster than using ggplot2, but not customizable
 #' @param assays A vector of assays to pull data from
+#' @param combine Combine plots into a single \code{\link[patchwork]{patchwork}ed}
+#' ggplot object. If \code{FALSE}, return a list of ggplot objects
 #'
-#' @return No return value by default. If using fast = FALSE, will return a ggplot object.
+#' @return No return value by default. If using fast = FALSE, will return a
+#' \code{\link[patchwork]{patchwork}ed} ggplot object if combine = TRUE, otherwise
+#' returns a list of ggplot objects
 #'
+#' @importFrom patchwork wrap_plots
 #' @export
 #'
 #' @seealso \code{\link[graphics]{image}} \code{\link[ggplot2]{geom_raster}}
@@ -43,11 +48,11 @@ DimHeatmap <- function(
   balanced = TRUE,
   projected = FALSE,
   ncol = NULL,
-  combine = TRUE,
   fast = TRUE,
   raster = TRUE,
   slot = 'scale.data',
-  assays = NULL
+  assays = NULL,
+  combine = TRUE
 ) {
   ncol <- ncol %||% ifelse(test = length(x = dims) > 2, yes = 3, no = length(x = dims))
   plots <- vector(mode = 'list', length = length(x = dims))
@@ -157,11 +162,7 @@ DimHeatmap <- function(
     return(invisible(x = NULL))
   }
   if (combine) {
-    plots <- CombinePlots(
-      plots = plots,
-      ncol = ncol,
-      legend = 'right'
-    )
+    plots <- wrap_plots(plots, ncol = ncol, guides = "collect")
   }
   return(plots)
 }
@@ -193,15 +194,17 @@ DimHeatmap <- function(
 #' @param lines.width Integer number to adjust the width of the separating white lines.
 #' Corresponds to the number of "cells" between each group.
 #' @param group.bar.height Scale the height of the color bar
-#' @param combine Combine plots into a single gg object; note that if TRUE; themeing will not work
-#' when plotting multiple dimensions
+#' @param combine Combine plots into a single \code{\link[patchwork]{patchwork}ed}
+#' ggplot object. If \code{FALSE}, return a list of ggplot objects
 #'
-#' @return A ggplot object
+#' @return A \code{\link[patchwork]{patchwork}ed} ggplot object if
+#' \code{combine = TRUE}; otherwise, a list of ggplot objects
 #'
 #' @importFrom stats median
 #' @importFrom scales hue_pal
 #' @importFrom ggplot2 annotation_raster coord_cartesian scale_color_manual
 #' ggplot_build aes_string
+#' @importFrom patchwork wrap_plots
 #' @export
 #'
 #' @examples
@@ -355,7 +358,8 @@ DoHeatmap <- function(
         scale_color_manual(values = cols)
       if (label) {
         x.max <- max(pbuild$layout$panel_params[[1]]$x.range)
-        x.divs <- pbuild$layout$panel_params[[1]]$x.major
+        # Attempt to pull xdivs from x.major in ggplot2 < 3.3.0; if NULL, pull from the >= 3.3.0 slot
+        x.divs <- pbuild$layout$panel_params[[1]]$x.major %||% pbuild$layout$panel_params[[1]]$x$break_positions()
         x <- data.frame(group = sort(x = group.use), x = x.divs)
         label.x.pos <- tapply(X = x$x, INDEX = x$group, FUN = median) * x.max
         label.x.pos <- data.frame(group = names(x = label.x.pos), label.x.pos)
@@ -378,7 +382,7 @@ DoHeatmap <- function(
     plots[[i]] <- plot
   }
   if (combine) {
-    plots <- CombinePlots(plots = plots)
+    plots <- wrap_plots(plots)
   }
   return(plots)
 }
@@ -469,11 +473,12 @@ HTOHeatmap <- function(
 #' @param same.y.lims Set all the y-axis limits to the same values
 #' @param log plot the feature axis on log scale
 #' @param ncol Number of columns if multiple plots are displayed
-#' @param combine Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features
 #' @param slot Use non-normalized counts data for plotting
-#' @param ... Extra parameters passed on to \code{\link{CombinePlots}}
+#' @param combine Combine plots into a single \code{\link[patchwork]{patchwork}ed}
+#' ggplot object. If \code{FALSE}, return a list of ggplot objects
 #'
-#' @return A ggplot object
+#' @return A \code{\link[patchwork]{patchwork}ed} ggplot object if
+#' \code{combine = TRUE}; otherwise, a list of ggplot objects
 #'
 #' @export
 #'
@@ -492,9 +497,8 @@ RidgePlot <- function(
   same.y.lims = FALSE,
   log = FALSE,
   ncol = NULL,
-  combine = TRUE,
   slot = 'data',
-  ...
+  combine = TRUE
 ) {
   return(ExIPlot(
     object = object,
@@ -509,9 +513,8 @@ RidgePlot <- function(
     cols = cols,
     group.by = group.by,
     log = log,
-    combine = combine,
     slot = slot,
-    ...
+    combine = combine
   ))
 }
 
@@ -523,11 +526,12 @@ RidgePlot <- function(
 #' @inheritParams RidgePlot
 #' @param pt.size Point size for geom_violin
 #' @param split.by A variable to split the violin plots by,
-#' @param multi.group  plot each group of the split violin plots by multiple or single violin shapes
-#' see \code{\link{FetchData}} for more details
+#' @param split.plot  plot each group of the split violin plots by multiple or
+#' single violin shapes.
 #' @param adjust Adjust parameter for geom_violin
 #'
-#' @return A ggplot object
+#' @return A \code{\link[patchwork]{patchwork}ed} ggplot object if
+#' \code{combine = TRUE}; otherwise, a list of ggplot objects
 #'
 #' @export
 #'
@@ -552,14 +556,26 @@ VlnPlot <- function(
   same.y.lims = FALSE,
   log = FALSE,
   ncol = NULL,
-  combine = TRUE,
   slot = 'data',
-  multi.group = FALSE,
-  ...
+  split.plot = FALSE,
+  combine = TRUE
 ) {
+  if (
+    !is.null(x = split.by) &
+    getOption(x = 'Seurat.warn.vlnplot.split', default = TRUE)
+  ) {
+    message(
+      "The default behaviour of split.by has changed.\n",
+      "Separate violin plots are now plotted side-by-side.\n",
+      "To restore the old behaviour of a single split violin,\n",
+      "set split.plot = TRUE.
+      \nThis message will be shown once per session."
+    )
+    options(Seurat.warn.vlnplot.split = FALSE)
+  }
   return(ExIPlot(
     object = object,
-    type = ifelse(test = multi.group, yes = 'multiViolin', no = 'violin'),
+    type = ifelse(test = split.plot, yes = 'splitViolin', no = 'violin'),
     features = features,
     idents = idents,
     ncol = ncol,
@@ -574,8 +590,7 @@ VlnPlot <- function(
     split.by = split.by,
     log = log,
     slot = slot,
-    combine = combine,
-    ...
+    combine = combine
   ))
 }
 
@@ -598,6 +613,8 @@ VlnPlot <- function(
 #' @return Returns a DimPlot
 #'
 #' @export
+#'
+#' @seealso \code{\link{DimPlot}}
 #'
 #' @examples
 #' pbmc_small
@@ -689,14 +706,16 @@ ColorDimSplit <- function(
 #' @param sizes.highlight Size of highlighted cells; will repeat to the length
 #' groups in cells.highlight
 #' @param na.value Color value for NA points when using custom scale
-#' @param combine Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features
 #' @param ncol Number of columns for display when combining plots
-#' @param ... Extra parameters passed on to \code{\link{CombinePlots}}
+#' @param combine Combine plots into a single \code{\link[patchwork]{patchwork}ed}
+#' ggplot object. If \code{FALSE}, return a list of ggplot objects
 #'
-#' @return A ggplot object
+#' @return A \code{\link[patchwork]{patchwork}ed} ggplot object if
+#' \code{combine = TRUE}; otherwise, a list of ggplot objects
 #'
 #' @importFrom rlang !!
 #' @importFrom ggplot2 facet_wrap vars sym
+#' @importFrom patchwork wrap_plots
 #'
 #' @export
 #'
@@ -705,7 +724,7 @@ ColorDimSplit <- function(
 #'
 #' @aliases TSNEPlot PCAPlot ICAPlot
 #' @seealso \code{\link{FeaturePlot}} \code{\link{HoverLocator}}
-#' \code{\link{CellSelector}} \code{link{FetchData}}
+#' \code{\link{CellSelector}} \code{\link{FetchData}}
 #'
 #' @examples
 #' DimPlot(object = pbmc_small)
@@ -729,11 +748,9 @@ DimPlot <- function(
   cols.highlight = '#DE2D26',
   sizes.highlight = 1,
   na.value = 'grey50',
-  combine = TRUE,
   ncol = NULL,
-  ...
+  combine = TRUE
 ) {
-  CheckDots(..., fxns = 'CombinePlots')
   if (length(x = dims) != 2) {
     stop("'dims' must be a two-length vector")
   }
@@ -743,6 +760,7 @@ DimPlot <- function(
   data <- as.data.frame(x = data)
   dims <- paste0(Key(object = object[[reduction]]), dims)
   object[['ident']] <- Idents(object = object)
+  orig.groups <- group.by
   group.by <- group.by %||% 'ident'
   data[, group.by] <- object[[group.by]][cells, , drop = FALSE]
   for (group in group.by) {
@@ -796,16 +814,11 @@ DimPlot <- function(
       return(plot)
     }
   )
+  if (!is.null(x = split.by)) {
+    ncol <- 1
+  }
   if (combine) {
-    plots <- CombinePlots(
-      plots = plots,
-      ncol = if (!is.null(x = split.by) && length(x = group.by) > 1) {
-        1
-      } else {
-        ncol
-      },
-      ...
-    )
+    plots <- wrap_plots(plots, ncol = orig.groups %iff% ncol)
   }
   return(plots)
 }
@@ -842,16 +855,18 @@ DimPlot <- function(
 #' @param blend Scale and blend expression values to visualize coexpression of two features
 #' @param blend.threshold The color cutoff from weak signal to strong signal; ranges from 0 to 1.
 #' @param ncol Number of columns to combine multiple feature plots to, ignored if \code{split.by} is not \code{NULL}
-#' @param combine Combine plots into a single gg object; note that if \code{TRUE}; themeing will not work when plotting multiple features
 #' @param coord.fixed Plot cartesian coordinates with fixed aspect ratio
 #' @param by.col If splitting by a factor, plot the splits per column with the features as rows; ignored if \code{blend = TRUE}
-#' @param sort.cell If \code{TRUE}, the positive cells will overlap the negative cells
+#' @param sort.cell Redundant with \code{order}. This argument is being
+#' deprecated. Please use \code{order} instead.
+#' @param combine Combine plots into a single \code{\link[patchwork]{patchwork}ed}
+#' ggplot object. If \code{FALSE}, return a list of ggplot objects
 #'
-#' @return Returns a ggplot object if only 1 feature is plotted.
-#' If >1 features are plotted and \code{combine=TRUE}, returns a combined ggplot object using \code{cowplot::plot_grid}.
-#' If >1 features are plotted and \code{combine=FALSE}, returns a list of ggplot objects.
+#' @return A \code{\link[patchwork]{patchwork}ed} ggplot object if
+#' \code{combine = TRUE}; otherwise, a list of ggplot objects
 #'
 #' @importFrom grDevices rgb
+#' @importFrom patchwork wrap_plots
 #' @importFrom cowplot theme_cowplot
 #' @importFrom RColorBrewer brewer.pal.info
 #' @importFrom ggplot2 labs scale_x_continuous scale_y_continuous theme element_rect
@@ -894,11 +909,23 @@ FeaturePlot <- function(
   label.size = 4,
   repel = FALSE,
   ncol = NULL,
-  combine = TRUE,
   coord.fixed = FALSE,
   by.col = TRUE,
-  sort.cell = FALSE
+  sort.cell = NULL,
+  combine = TRUE
 ) {
+  # TODO: deprecate fully on 3.2.0
+  if (!is.null(x = sort.cell)) {
+    warning(
+      "The sort.cell parameter is being deprecated. Please use the order ",
+      "parameter instead for equivalent functionality.",
+      call. = FALSE,
+      immediate. = TRUE
+    )
+    if (isTRUE(x = sort.cell)) {
+      order <- sort.cell
+    }
+  }
   # Set a theme to remove right-hand Y axis lines
   # Also sets right-hand Y axis text label formatting
   no.right <- theme(
@@ -1124,9 +1151,6 @@ FeaturePlot <- function(
         cols.use <- NULL
       }
       data.single <- data.plot[, c(dims, 'ident', feature, shape.by)]
-      if (sort.cell) {
-        data.single <- data.single[order(data.single[, feature]),]
-      }
       # Make the plot
       plot <- SingleDimPlot(
         data = data.single,
@@ -1164,7 +1188,10 @@ FeaturePlot <- function(
         if (j == length(x = features) && !blend) {
           suppressMessages(
             expr = plot <- plot +
-              scale_y_continuous(sec.axis = dup_axis(name = ident)) +
+              scale_y_continuous(
+                sec.axis = dup_axis(name = ident),
+                limits = ylims
+              ) +
               no.right
           )
         }
@@ -1258,30 +1285,30 @@ FeaturePlot <- function(
   # Remove NULL plots
   plots <- Filter(f = Negate(f = is.null), x = plots)
   # Combine the plots
+  if (is.null(x = ncol)) {
+    ncol <- 2
+    if (length(x = features) == 1) {
+      ncol <- 1
+    }
+    if (length(x = features) > 6) {
+      ncol <- 3
+    }
+    if (length(x = features) > 9) {
+      ncol <- 4
+    }
+  }
+  ncol <- ifelse(
+    test = is.null(x = split.by) || blend,
+    yes = ncol,
+    no = length(x = features)
+  )
+  legend <- if (blend) {
+    'none'
+  } else {
+    split.by %iff% 'none'
+  }
+  # Transpose the FeatureHeatmap matrix (not applicable for blended FeaturePlots)
   if (combine) {
-    if (is.null(x = ncol)) {
-      ncol <- 2
-      if (length(x = features) == 1) {
-        ncol <- 1
-      }
-      if (length(x = features) > 6) {
-        ncol <- 3
-      }
-      if (length(x = features) > 9) {
-        ncol <- 4
-      }
-    }
-    ncol <- ifelse(
-      test = is.null(x = split.by) || blend,
-      yes = ncol,
-      no = length(x = features)
-    )
-    legend <- if (blend) {
-      'none'
-    } else {
-      split.by %iff% 'none'
-    }
-    # Transpose the FeatureHeatmap matrix (not applicable for blended FeaturePlots)
     if (by.col && !is.null(x = split.by) && !blend) {
       plots <- lapply(
         X = plots,
@@ -1290,7 +1317,7 @@ FeaturePlot <- function(
             expr = x +
               theme_cowplot() +
               ggtitle("") +
-              scale_y_continuous(sec.axis = dup_axis(name = "")) +
+              scale_y_continuous(sec.axis = dup_axis(name = ""), limits = ylims) +
               no.right
           ))
         }
@@ -1298,37 +1325,50 @@ FeaturePlot <- function(
       nsplits <- length(x = levels(x = data$split))
       idx <- 1
       for (i in (length(x = features) * (nsplits - 1) + 1):(length(x = features) * nsplits)) {
-        plots[[i]] <- suppressMessages(plots[[i]] + scale_y_continuous(sec.axis = dup_axis(name = features[[idx]])) + no.right)
+        plots[[i]] <- suppressMessages(
+          expr = plots[[i]] +
+            scale_y_continuous(
+              sec.axis = dup_axis(name = features[[idx]]),
+              limits = ylims
+            ) +
+            no.right
+        )
         idx <- idx + 1
       }
       idx <- 1
       for (i in which(x = 1:length(x = plots) %% length(x = features) == 1)) {
-        plots[[i]] <- plots[[i]] + ggtitle(levels(x = data$split)[[idx]]) + theme(plot.title = element_text(hjust = 0.5))
+        plots[[i]] <- plots[[i]] +
+          ggtitle(levels(x = data$split)[[idx]]) +
+          theme(plot.title = element_text(hjust = 0.5))
         idx <- idx + 1
       }
       idx <- 1
       if (length(x = features) == 1) {
         for (i in 1:length(x = plots)) {
-          plots[[i]] <- plots[[i]] + ggtitle(levels(x = data$split)[[idx]]) + theme(plot.title = element_text(hjust = 0.5))
+          plots[[i]] <- plots[[i]] +
+            ggtitle(levels(x = data$split)[[idx]]) +
+            theme(plot.title = element_text(hjust = 0.5))
           idx <- idx + 1
         }
+        ncol <- 1
+        nrow <- nsplits
+      } else {
+        nrow <- split.by %iff% length(x = levels(x = data$split))
       }
       plots <- plots[c(do.call(
         what = rbind,
-        args = split(x = 1:length(x = plots), f = ceiling(x = seq_along(along.with = 1:length(x = plots))/length(x = features)))
+        args = split(x = 1:length(x = plots), f = ceiling(x = seq_along(along.with = 1:length(x = plots)) / length(x = features)))
       ))]
-      plots <- CombinePlots(
-        plots = plots,
-        ncol = nsplits,
-        legend = legend
-      )
+      # Set ncol to number of splits (nrow) and nrow to number of features (ncol)
+      plots <- wrap_plots(plots, ncol = nrow, nrow = ncol)
+      if (!is.null(x = legend) && legend == 'none') {
+        plots <- plots & NoLegend()
+      }
     } else {
-      plots <- CombinePlots(
-        plots = plots,
-        ncol = ncol,
-        legend = legend,
-        nrow = split.by %iff% length(x = levels(x = data$split))
-      )
+      plots <- wrap_plots(plots, ncol = ncol, nrow = split.by %iff% length(x = levels(x = data$split)))
+    }
+    if (!is.null(x = legend) && legend == 'none') {
+      plots <- plots & NoLegend()
     }
   }
   return(plots)
@@ -1403,10 +1443,13 @@ CellScatter <- function(
 #' @param span Spline span in loess function call, if \code{NULL}, no spline added
 #' @param smooth Smooth the graph (similar to smoothScatter)
 #' @param slot Slot to pull data from, should be one of 'counts', 'data', or 'scale.data'
+#' @param combine Combine plots into a single \code{\link[patchwork]{patchwork}ed}
 #'
 #' @return A ggplot object
 #'
 #' @importFrom ggplot2 geom_smooth aes_string
+#' @importFrom patchwork wrap_plots
+#'
 #' @export
 #'
 #' @aliases GenePlot
@@ -1425,28 +1468,51 @@ FeatureScatter <- function(
   shape.by = NULL,
   span = NULL,
   smooth = FALSE,
+  combine = TRUE,
   slot = 'data'
 ) {
   cells <- cells %||% colnames(x = object)
-  group.by <- group.by %||% Idents(object = object)[cells]
-  if (length(x = group.by) == 1) {
-    group.by <- object[[]][, group.by]
-  }
-  plot <- SingleCorPlot(
-    data = FetchData(
-      object = object,
-      vars = c(feature1, feature2),
-      cells = cells,
-      slot = slot
-    ),
-    col.by = group.by,
-    cols = cols,
-    pt.size = pt.size,
-    smooth = smooth,
-    legend.title = 'Identity',
-    span = span
+  object[['ident']] <- Idents(object = object)
+  group.by <- group.by %||% 'ident'
+  data <-  FetchData(
+    object = object,
+    vars = c(feature1, feature2, group.by),
+    cells = cells,
+    slot = slot
   )
-  return(plot)
+  if (isFALSE(x = feature1 %in% colnames(x = data))) {
+    stop("Feature 1 (", feature1, ") not found.", call. = FALSE)
+  }
+  if (isFALSE(x = feature2 %in% colnames(x = data))) {
+    stop("Feature 2 (", feature2, ") not found.", call. = FALSE)
+  }
+  data <- as.data.frame(x = data)
+  for (group in group.by) {
+    if (!is.factor(x = data[, group])) {
+      data[, group] <- factor(x = data[, group])
+    }
+  }
+  plots <- lapply(
+    X = group.by,
+    FUN = function(x) {
+      SingleCorPlot(
+        data = data[,c(feature1, feature2)],
+        col.by = data[, x],
+        cols = cols,
+        pt.size = pt.size,
+        smooth = smooth,
+        legend.title = 'Identity',
+        span = span
+      )
+    }
+  )
+  if (isTRUE(x = length(x = plots) == 1)) {
+    return(plots[[1]])
+  }
+  if (isTRUE(x = combine)) {
+    plots <- wrap_plots(plots, ncol = length(x = group.by))
+  }
+  return(plots)
 }
 
 #' View variable features
@@ -1686,11 +1752,12 @@ PolyFeaturePlot <- function(
 #' The transition from "signal" to "noise" in the is hard to see because the
 #' first singular value spacings are so large. Nicer visualizations result from
 #' skipping the first few. If set to 0 (default) starts from k/2.
-#' @param combine Combine plots into a single gg object; note that if TRUE,
-#' themeing will not work when plotting multiple features
+#' @param combine Combine plots into a single \code{\link[patchwork]{patchwork}ed}
+#' ggplot object. If \code{FALSE}, return a list of ggplot objects
 #'
-#' @return A list of 3 ggplot objects splotting the singular values, the
-#' spacings of the singular values, and the p-values of the singular values.
+#' @return A list of 3 \code{\link[patchwork]{patchwork}ed} ggplot objects
+#' splotting the singular values, the spacings of the singular values, and the
+#' p-values of the singular values.
 #'
 #' @author Jun Zhao, George Linderman
 #' @seealso \code{\link{RunALRA}}
@@ -1698,6 +1765,7 @@ PolyFeaturePlot <- function(
 #' @importFrom cowplot theme_cowplot
 #' @importFrom ggplot2 ggplot aes_string geom_point geom_line
 #' geom_vline scale_x_continuous labs
+#' @importFrom patchwork wrap_plots
 #' @export
 #'
 ALRAChooseKPlot <- function(object, start = 0, combine = TRUE) {
@@ -1740,7 +1808,7 @@ ALRAChooseKPlot <- function(object, start = 0, combine = TRUE) {
     labs(x = NULL, y = 's_{i} - s_{i-1}', title = 'Singular value spacings')
   plots <- list(spectrum = gg1, spacings = gg2)
   if (combine) {
-    plots <- CombinePlots(plots = plots)
+    plots <- wrap_plots(plots)
   }
   return(plots)
 }
@@ -1842,6 +1910,7 @@ BarcodeInflectionsPlot <- function(object) {
 #' @param group.by Factor to group the cells by
 #' @param split.by Factor to split the groups by (replicates the functionality of the old SplitDotPlotGG);
 #' see \code{\link{FetchData}} for more details
+#' @param scale Determine whether the data is scaled, TRUE for default
 #' @param scale.by Scale the size of the points by 'size' or by 'radius'
 #' @param scale.min Set lower limit for scaling, use NA for default
 #' @param scale.max Set upper limit for scaling, use NA for default
@@ -1874,6 +1943,7 @@ DotPlot <- function(
   dot.scale = 6,
   group.by = NULL,
   split.by = NULL,
+  scale = TRUE,
   scale.by = 'radius',
   scale.min = NA,
   scale.max = NA
@@ -1937,15 +2007,26 @@ DotPlot <- function(
   if (!is.null(x = id.levels)) {
     data.plot$id <- factor(x = data.plot$id, levels = id.levels)
   }
-  avg.exp.scaled <- sapply(
-    X = unique(x = data.plot$features.plot),
-    FUN = function(x) {
-      data.use <- data.plot[data.plot$features.plot == x, 'avg.exp']
-      data.use <- scale(x = data.use)
-      data.use <- MinMax(data = data.use, min = col.min, max = col.max)
-      return(data.use)
-    }
-  )
+ if (length(x = levels(x = data.plot$id)) == 1) {
+   scale <- FALSE
+   warning("Only one identity present, the expression values will be not scaled.")
+ }
+    avg.exp.scaled <- sapply(
+      X = unique(x = data.plot$features.plot),
+      FUN = function(x) {
+        data.use <- data.plot[data.plot$features.plot == x, 'avg.exp']
+        if (scale) {
+          data.use <- scale(x = data.use)
+          data.use <- MinMax(data = data.use, min = col.min, max = col.max)
+        } else {
+          data.use <- log(x = data.use)
+        }
+        return(data.use)
+      }
+    )
+
+
+
   avg.exp.scaled <- as.vector(x = t(x = avg.exp.scaled))
   if (!is.null(x = split.by)) {
     avg.exp.scaled <- as.numeric(x = cut(x = avg.exp.scaled, breaks = 20))
@@ -1959,10 +2040,16 @@ DotPlot <- function(
   data.plot$pct.exp <- data.plot$pct.exp * 100
   if (!is.null(x = split.by)) {
     splits.use <- vapply(
-      X = strsplit(x = as.character(x = data.plot$id), split = '_'),
-      FUN = '[[',
+      X = as.character(x = data.plot$id),
+      FUN = gsub,
       FUN.VALUE = character(length = 1L),
-      2
+      pattern =  paste0(
+        '^((',
+        paste(sort(x = levels(x = object), decreasing = TRUE), collapse = '|'),
+        ')_)'
+      ),
+      replacement = '',
+      USE.NAMES = FALSE
     )
     data.plot$colors <- mapply(
       FUN = function(color, value) {
@@ -2164,11 +2251,13 @@ PlotClusterTree <- function(object, ...) {
 #' @param balanced Return an equal number of genes with + and - scores. If
 #' FALSE (default), returns the top genes ranked by the scores absolute values
 #' @param ncol Number of columns to display
-#' @param combine Combine plots into a single gg object; note that if TRUE;
-#' themeing will not work when plotting multiple features
+#' @param combine Combine plots into a single \code{\link[patchwork]{patchwork}ed}
+#' ggplot object. If \code{FALSE}, return a list of ggplot objects
 #'
-#' @return A ggplot object
+#' @return A \code{\link[patchwork]{patchwork}ed} ggplot object if
+#' \code{combine = TRUE}; otherwise, a list of ggplot objects
 #'
+#' @importFrom patchwork wrap_plots
 #' @importFrom cowplot theme_cowplot
 #' @importFrom ggplot2 ggplot aes_string geom_point labs
 #' @export
@@ -2187,17 +2276,15 @@ VizDimLoadings <- function(
   ncol = NULL,
   combine = TRUE
 ) {
-  if (is.null(x = ncol)) {
-    ncol <- 2
-    if (length(x = dims) == 1) {
-      ncol <- 1
-    }
-    if (length(x = dims) > 6) {
-      ncol <- 3
-    }
-    if (length(x = dims) > 9) {
-      ncol <- 4
-    }
+  ncol <- ncol %||% 2
+  if (length(x = dims) == 1) {
+    ncol <- 1
+  }
+  if (length(x = dims) > 6) {
+    ncol <- 3
+  }
+  if (length(x = dims) > 9) {
+    ncol <- 4
   }
   loadings <- Loadings(object = object[[reduction]], projected = projected)
   features <- lapply(
@@ -2231,7 +2318,7 @@ VizDimLoadings <- function(
     }
   )
   if (combine) {
-    plots <- CombinePlots(plots = plots, ncol = ncol, legend = NULL)
+    plots <- wrap_plots(plots, ncol = ncol)
   }
   return(plots)
 }
@@ -2415,7 +2502,7 @@ CollapseEmbeddingOutliers <- function(
   data.medians.scale <- as.matrix(x = scale(x = data.medians, center = TRUE, scale = TRUE))
   data.medians.scale[abs(x = data.medians.scale) < outlier.sd] <- 0
   data.medians.scale <- sign(x = data.medians.scale) * (abs(x = data.medians.scale) - outlier.sd)
-  data.correct <- sweep(
+  data.correct <- Sweep(
     x = data.medians.scale,
     MARGIN = 2,
     STATS = data.sd,
@@ -2425,7 +2512,7 @@ CollapseEmbeddingOutliers <- function(
   new.embeddings <- embeddings
   for (i in rownames(x = data.correct)) {
     cells.correct <- rownames(x = idents)[idents[, "ident"] == i]
-    new.embeddings[cells.correct, ] <- sweep(
+    new.embeddings[cells.correct, ] <- Sweep(
       x = new.embeddings[cells.correct,],
       MARGIN = 2,
       STATS = data.correct[i, ],
@@ -2461,19 +2548,24 @@ CollapseEmbeddingOutliers <- function(
 #'   size = ncol(x = pbmc_small),
 #'   replace = TRUE
 #' )
-#' plots <- FeaturePlot(
+#' plot1 <- FeaturePlot(
 #'   object = pbmc_small,
-#'   features = c('MS4A1', 'FCN1'),
-#'   split.by = 'group',
-#'   combine = FALSE
+#'   features = 'MS4A1',
+#'   split.by = 'group'
+#' )
+#' plot2 <- FeaturePlot(
+#'   object = pbmc_small,
+#'   features = 'FCN1',
+#'   split.by = 'group'
 #' )
 #' CombinePlots(
-#'   plots = plots,
+#'   plots = list(plot1, plot2),
 #'   legend = 'none',
 #'   nrow = length(x = unique(x = pbmc_small[['group', drop = TRUE]]))
 #' )
 #'
 CombinePlots <- function(plots, ncol = NULL, legend = NULL, ...) {
+  .Deprecated(msg = "CombinePlots is being deprecated. Plots should now be combined using the patchwork system.")
   plots.combined <- if (length(x = plots) > 1) {
     if (!is.null(x = legend)) {
       if (legend != 'none') {
@@ -2586,7 +2678,7 @@ CustomPalette <- function(
 #'
 #' @details
 #' Taken from the pals package (Licence: GPL-3).
-#' \url{https://cran.r-project.org/web/packages/pals/index.html}
+#' \url{https://cran.r-project.org/package=pals}
 #' Credit: Kevin Wright
 #'
 #' @export
@@ -3425,7 +3517,7 @@ BlendMap <- function(color.matrix) {
 #
 # @return An n x n matrix of blended colors
 #
-#' @importFrom grDevices rgb colorRamp
+#' @importFrom grDevices col2rgb
 #
 BlendMatrix <- function(
   n = 10,
@@ -3436,10 +3528,13 @@ BlendMatrix <- function(
   if (0 > col.threshold || col.threshold > 1) {
     stop("col.threshold must be between 0 and 1")
   }
-  C0 <- colorRamp(colors = negative.color)(1)
-  ramp <- colorRamp(colors = two.colors)
-  C1 <- ramp(x = 0)
-  C2 <- ramp(x = 1)
+  C0 <- as.vector(col2rgb(negative.color, alpha = TRUE))
+  C1 <- as.vector(col2rgb(two.colors[1], alpha = TRUE))
+  C2 <- as.vector(col2rgb(two.colors[2], alpha = TRUE))
+  blend_alpha <- (C1[4] + C2[4])/2
+  C0 <- C0[-4]
+  C1 <- C1[-4]
+  C2 <- C2[-4]
   merge.weight <- min(255 / (C1 + C2 +  C0 + 0.01))
   sigmoid <- function(x) {
     return(1 / (1 + exp(-x)))
@@ -3452,6 +3547,7 @@ BlendMatrix <- function(
     C0,
     C1,
     C2,
+    alpha,
     merge.weight
   ) {
     c.min <- sigmoid(5 * (1 / n - col.threshold))
@@ -3472,10 +3568,10 @@ BlendMatrix <- function(
     C_blend[C_blend > 255] <- 255
     C_blend[C_blend < 0] <- 0
     return(rgb(
-      red = C_blend[, 1],
-      green = C_blend[, 2],
-      blue = C_blend[, 3],
-      alpha = 255,
+      red = C_blend[1],
+      green = C_blend[2],
+      blue = C_blend[3],
+      alpha = alpha,
       maxColorValue = 255
     ))
   }
@@ -3490,6 +3586,7 @@ BlendMatrix <- function(
         C0 = C0,
         C1 = C1,
         C2 = C2,
+        alpha = blend_alpha,
         merge.weight = merge.weight
       )
     }
@@ -3661,7 +3758,7 @@ GuideProteinPlot <- function(object,
 # Basically combines the codebase for VlnPlot and RidgePlot
 #
 # @param object Seurat object
-# @param type Plot type, choose from 'ridge', 'violin', or 'multiViolin'
+# @param type Plot type, choose from 'ridge', 'violin', or 'splitViolin'
 # @param features Features to plot (gene expression, metrics, PC scores,
 # anything that can be retreived by FetchData)
 # @param idents Which classes to include in the plot (default is all)
@@ -3675,12 +3772,16 @@ GuideProteinPlot <- function(object,
 # @param group.by Group (color) cells in different ways (for example, orig.ident)
 # @param split.by A variable to split the plot by
 # @param log plot Y axis on log scale
-# @param combine Combine plots using cowplot::plot_grid
 # @param slot Use non-normalized counts data for plotting
-# @param ... Extra parameters passed to \code{\link{CombinePlots}}
+# @param combine Combine plots into a single \code{\link[patchwork]{patchwork}ed}
+# ggplot object. If \code{FALSE}, return a list of ggplot objects
+#
+# @return A \code{\link[patchwork]{patchwork}ed} ggplot object if
+# \code{combine = TRUE}; otherwise, a list of ggplot objects
 #
 #' @importFrom scales hue_pal
 #' @importFrom ggplot2 xlab ylab
+#' @importFrom patchwork wrap_plots
 #
 ExIPlot <- function(
   object,
@@ -3698,9 +3799,8 @@ ExIPlot <- function(
   group.by = NULL,
   split.by = NULL,
   log = FALSE,
-  combine = TRUE,
   slot = 'data',
-  ...
+  combine = TRUE
 ) {
   assay <- assay %||% DefaultAssay(object = object)
   DefaultAssay(object = object) <- assay
@@ -3746,6 +3846,10 @@ ExIPlot <- function(
     }
     cols <- rep_len(x = cols, length.out = length(x = levels(x = split)))
     names(x = cols) <- sort(x = levels(x = split))
+    if ((length(x = cols) > 2) & (type == "splitViolin")) {
+      warning("Split violin is only supported for <3 groups, using multi-violin.")
+      type <- "violin"
+    }
   }
   if (same.y.lims && is.null(x = y.max)) {
     y.max <- max(data)
@@ -3763,15 +3867,14 @@ ExIPlot <- function(
         adjust = adjust,
         cols = cols,
         pt.size = pt.size,
-        log = log,
-        ...
+        log = log
       ))
     }
   )
   label.fxn <- switch(
     EXPR = type,
     'violin' = ylab,
-    "multiViolin" = ylab,
+    "splitViolin" = ylab,
     'ridge' = xlab,
     stop("Unknown ExIPlot type ", type, call. = FALSE)
   )
@@ -3792,15 +3895,10 @@ ExIPlot <- function(
     }
   }
   if (combine) {
-    combine.args <- list(
-      'plots' = plots,
-      'ncol' = ncol
-    )
-    combine.args <- c(combine.args, list(...))
-    if (!'legend' %in% names(x = combine.args)) {
-      combine.args$legend <- 'none'
+    plots <- wrap_plots(plots, ncol = ncol)
+    if (length(x = features) > 1) {
+      plots <- plots & NoLegend()
     }
-    plots <- do.call(what = 'CombinePlots', args = combine.args)
   }
   return(plots)
 }
@@ -4127,10 +4225,11 @@ PlotBuild <- function(data, dark.theme = FALSE, smooth = FALSE, ...) {
 # @return A dataframe of x and y coordinates for points selected
 #
 #' @importFrom graphics locator
-#' @importFrom SDMTools pnt.in.poly
+# @importFrom SDMTools pnt.in.poly
 #
 PointLocator <- function(plot, recolor = TRUE, dark.theme = FALSE, ...) {
   #   Convert the ggplot object to a data.frame
+  PackageCheck('SDMTools')
   plot.data <- GGpointToBase(plot = plot, dark.theme = dark.theme, ...)
   npoints <- nrow(x = plot.data)
   cat("Click around the cluster of points you wish to select\n")
@@ -4139,7 +4238,7 @@ PointLocator <- function(plot, recolor = TRUE, dark.theme = FALSE, ...) {
   polygon <- locator(n = npoints, type = 'l')
   polygon <- data.frame(polygon)
   #   pnt.in.poly returns a data.frame of points
-  points.all <- pnt.in.poly(
+  points.all <- SDMTools::pnt.in.poly(
     pnts = plot.data[, c(1, 2)],
     poly.pnts = polygon
   )
@@ -4615,7 +4714,7 @@ SingleDimPlot <- function(
 #' @importFrom stats rnorm
 #' @importFrom utils globalVariables
 #' @importFrom ggridges geom_density_ridges theme_ridges
-#' @importFrom ggplot2 ggplot aes_string theme labs geom_violin geom_jitter ylim
+#' @importFrom ggplot2 ggplot aes_string theme labs geom_violin geom_jitter ylim position_jitterdodge
 #' scale_fill_manual scale_y_log10 scale_x_log10 scale_y_discrete scale_x_continuous waiver
 #' @importFrom cowplot theme_cowplot
 #'
@@ -4665,14 +4764,14 @@ SingleExIPlot <- function(
     data[, feature] <- data[, feature] + noise
   }
   axis.label <- 'Expression Level'
-  y.max <- y.max %||% max(data[, feature])
+  y.max <- y.max %||% max(data[, feature][is.finite(x = data[, feature])])
   if (type == 'violin' && !is.null(x = split)) {
     data$split <- split
-    vln.geom <- geom_split_violin
-    fill <- 'split'
-  } else if (type == 'multiViolin' && !is.null(x = split )) {
-    data$split <- split
     vln.geom <- geom_violin
+    fill <- 'split'
+  } else if (type == 'splitViolin' && !is.null(x = split )) {
+    data$split <- split
+    vln.geom <- geom_split_violin
     fill <- 'split'
     type <- 'violin'
   } else {
@@ -4690,7 +4789,14 @@ SingleExIPlot <- function(
         vln.geom(scale = 'width', adjust = adjust, trim = TRUE),
         theme(axis.text.x = element_text(angle = 45, hjust = 1))
       )
-      jitter <- geom_jitter(height = 0, size = pt.size)
+      if (is.null(x = split)) {
+        jitter <- geom_jitter(height = 0, size = pt.size)
+      } else {
+        jitter <- geom_jitter(
+          position = position_jitterdodge(jitter.width = 0.4, dodge.width = 0.9),
+          size = pt.size
+        )
+      }
       log.scale <- scale_y_log10()
       axis.scale <- ylim
     },